+++ frontend/src/lib/components/ResourcePicker.svelte
@@ -10,7 +10,7 @@
 	import { Pen, Plus, RotateCw } from 'lucide-svelte'
 	import { sendUserToast } from '$lib/toast'
 	import { isDbType } from './apps/components/display/dbtable/utils'
-	import Select from './Select.svelte'
+	import Select from './select/Select.svelte'
 	import { createDispatcherIfMounted } from '$lib/createDispatcherIfMounted'
 
 	const dispatch = createEventDispatcher()
