+++ frontend/src/lib/components/WorkerGroup.svelte
@@ -20,7 +20,7 @@
 	import { defaultTags, nativeTags, type AutoscalingConfig } from './worker_group'
 	import AutoscalingConfigEditor from './AutoscalingConfigEditor.svelte'
 	import TagsToListenTo from './TagsToListenTo.svelte'
-	import Select from './Select.svelte'
+	import Select from './select/Select.svelte'
 
 	function computeVCpuAndMemory(workers: [string, WorkerPing[]][]) {
 		let vcpus = 0
