+++ frontend/src/lib/components/auditLogs/AuditLogsFilters.svelte
@@ -34,7 +34,7 @@
 	import { onDestroy, tick, untrack } from 'svelte'
 	import ToggleButtonGroup from '../common/toggleButton-v2/ToggleButtonGroup.svelte'
 	import ToggleButton from '../common/toggleButton-v2/ToggleButton.svelte'
-	import Select from '../Select.svelte'
+	import Select from '../select/Select.svelte'
 	import { usePromise } from '$lib/svelte5Utils.svelte'
 
 	let usernames: string[] | undefined = $state()
